requirements.txt
aiobotocore==2.1.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
attrs==21.4.0
boto3==1.21.29
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
fonttools==4.30.0
frozenlist==1.3.0
fsspec==2022.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
minio==7.1.5
multidict==6.0.2
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.3.15
s3fs==2022.2.0
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.32
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
yarl==1.7.2
zipp==3.7.0
